Move init to evaluate.py, create new empty init
